+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.4.0 (in development)
+-----------------------
+- Added an `rst2json.core.get_docutils_config_files()` function for fetching
+  the list of config files Docutils will read from based on the environment
+
 v0.3.0 (2020-07-14)
 -------------------
 - **Breaking**: When not a string, the `format` argument to the `rst2json()`